requirements.txt
langchain
streamlit 
python-dotenv
openai
langchain_community
crewai
"crewai[tools]"
base-64